requirements.txt
pyperclip
keyboard
pyautogui
SpeechRecognition
sounddevice
pygetwindow
tk
pyaudio
wave
openai
openai-whisper